with heroID fixes and leveldata added

+++ FinalProject2D/Assets/Scripts/LevelData.cs
@@ -12,7 +12,7 @@ public class LevelData : ScriptableObject
     private int[] coinsReward = {400,600,800};
 
     public int getlevelId() => _levelId;
-    public float getCoinsReward(int stars) => coinsReward[stars-1];
+    public int getCoinsReward(int stars) => coinsReward[stars-1];
 
 
 }
